moved redis and pg config in a grouped local config folder
